+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 
 - [plugin] added support for `DocumentSymbolProviderMetadata` [#10811](https://github.com/eclipse-theia/theia/pull/10811) - Contributed on behalf of STMicroelectronics
 
+<a name="breaking_changes_1.24.0">[Breaking Changes:](#breaking_changes_1.24.0)</a>
+
+- [markers] `getOverlayIconColor` renamed to `getSeverityColor` [#10273](https://github.com/eclipse-theia/theia/pull/10273)
+- [markers] removed method `getOverlayIcon` - marker decorations now use `tailDecorations` instead [#10273](https://github.com/eclipse-theia/theia/pull/10273)
+
 ## v1.23.0 - 2/24/2022
 
 [1.23.0 Milestone](https://github.com/eclipse-theia/theia/milestone/31)